streamlit/requirements.txt
streamlit
onnxruntime
Pillow
numpy
requests
tqdm